requirements.txt
Flask==2.3.2
Pillow==10.0.0
Pillow==10.0.1
reportlab==4.0.4
